requirements.txt
amqp==5.0.6
asgiref==3.4.1
bcrypt==3.2.0
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
dnspython==2.1.0
email-validator==1.1.3
fastapi==0.68.1
greenlet==1.1.1
h11==0.12.0
idna==3.2
kombu==5.1.0
prompt-toolkit==3.0.20
psycopg2-binary==2.9.1
pycparser==2.20
pydantic==1.8.2
PyJWT==2.1.0
python-http-client==3.3.2
pytz==2021.1
requests==2.26.0
sendgrid==6.8.1
six==1.16.0
SQLAlchemy==1.4.23
starkbank-ecdsa==1.1.1
starlette==0.14.2
typing-extensions==3.10.0.2
urllib3==1.26.6
uvicorn==0.15.0
vine==5.0.0
wcwidth==0.2.5
